EcsClusterProvider creation
- Moved logic from EcsServerClusterProvider to EcsClusterProvider
