+++ clouddriver-ecs/src/main/java/com/netflix/spinnaker/clouddriver/ecs/controllers/EcsClusterController.java
@@ -18,7 +18,7 @@
 
 package com.netflix.spinnaker.clouddriver.ecs.controllers;
 
-import com.netflix.spinnaker.clouddriver.ecs.provider.view.EcsServerClusterProvider;
+import com.netflix.spinnaker.clouddriver.ecs.provider.view.EcsClusterProvider;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.RequestMapping;
@@ -29,18 +29,18 @@ import java.util.List;
 @RestController
 public class EcsClusterController {
 
-  EcsServerClusterProvider ecsServerClusterProvider;
+  EcsClusterProvider ecsClusterProvider;
 
   @Autowired
-  public EcsClusterController(EcsServerClusterProvider ecsServerClusterProvider) {
-    this.ecsServerClusterProvider = ecsServerClusterProvider;
+  public EcsClusterController(EcsClusterProvider ecsClusterProvider) {
+    this.ecsClusterProvider = ecsClusterProvider;
   }
 
 
   @RequestMapping(value = "/ecs/{account}/{region}/ecscluster")
   public List<String> findEcsClusters(@PathVariable("account") String account,
                                       @PathVariable("region") String region) {
-    return ecsServerClusterProvider.getEcsClusters(account, region);
+    return ecsClusterProvider.getEcsClusters(account, region);
   }
 
 }
