+++ clouddriver-ecs/src/main/java/com/netflix/spinnaker/clouddriver/ecs/provider/view/EcsClusterProvider.java
@@ -1,0 +1,73 @@
+/*
+ *
+ *  * Copyright 2017 Lookout, Inc.
+ *  *
+ *  * Licensed under the Apache License, Version 2.0 (the "License");
+ *  * you may not use this file except in compliance with the License.
+ *  * You may obtain a copy of the License at
+ *  *
+ *  *    http://www.apache.org/licenses/LICENSE-2.0
+ *  *
+ *  * Unless required by applicable law or agreed to in writing, software
+ *  * distributed under the License is distributed on an "AS IS" BASIS,
+ *  * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ *  * See the License for the specific language governing permissions and
+ *  * limitations under the License.
+ *
+ */
+
+package com.netflix.spinnaker.clouddriver.ecs.provider.view;
+
+import com.amazonaws.services.ecs.AmazonECS;
+import com.amazonaws.services.ecs.model.ListClustersRequest;
+import com.amazonaws.services.ecs.model.ListClustersResult;
+import com.netflix.spinnaker.clouddriver.aws.security.AmazonClientProvider;
+import com.netflix.spinnaker.clouddriver.aws.security.AmazonCredentials;
+import com.netflix.spinnaker.clouddriver.ecs.services.ContainerInformationService;
+import com.netflix.spinnaker.clouddriver.security.AccountCredentials;
+import com.netflix.spinnaker.clouddriver.security.AccountCredentialsProvider;
+import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.stereotype.Component;
+
+import java.util.ArrayList;
+import java.util.List;
+
+@Component
+public class EcsClusterProvider {
+
+  private AccountCredentialsProvider accountCredentialsProvider;
+  private AmazonClientProvider amazonClientProvider;
+
+  @Autowired
+  public EcsClusterProvider(AccountCredentialsProvider accountCredentialsProvider,
+                            AmazonClientProvider amazonClientProvider) {
+    this.accountCredentialsProvider = accountCredentialsProvider;
+    this.amazonClientProvider = amazonClientProvider;
+  }
+
+  public List<String> getEcsClusters(String account, String region) {
+
+    List<String> listCluster = new ArrayList<>();
+
+    for (AccountCredentials credentials: accountCredentialsProvider.getAll()) {
+      if (credentials instanceof AmazonCredentials) {
+        AmazonECS amazonECS = amazonClientProvider.getAmazonEcs(account,
+          ((AmazonCredentials) credentials).getCredentialsProvider(),
+          region);
+        ListClustersRequest listClustersRequest = new ListClustersRequest();
+        ListClustersResult listClustersResult = amazonECS.listClusters(listClustersRequest);
+
+        for (String clusterArn : listClustersResult.getClusterArns()) {
+          String ecsClusterName = inferClusterNameFromClusterArn(clusterArn);
+          listCluster.add(ecsClusterName);
+        }
+      }
+    }
+    return listCluster;
+  }
+
+  private String inferClusterNameFromClusterArn(String clusterArn) {
+    return clusterArn.split("/")[1];
+  }
+
+}
